mcp_servers/requirements.txt
fastapi
pydantic
pyyaml
uvicorn[standard]
python-dotenv